requirements.txt
django
django-widget-tweaks
django-phonenumber-field[phonenumbers]
